# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AD Admin Sample API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List Products",
                "description": "Display a listing of the Products.",
                "operationId": "List Products",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "Skip prodcuts",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "integer": {
                                "summary": "An integer value",
                                "value": "1"
                            }
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Limit prodcuts",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "integer": {
                                "summary": "An integer value",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Message"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show Product",
                "description": "Display detailed information for product.",
                "operationId": "Show Product",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "Product ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "integer": {
                                "summary": "An integer value",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Message"
                    }
                }
            }
        },
        "/api/rabbitmq/{channel}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get a message by channel",
                "description": "Consumes a message from the specified channel",
                "operationId": "getMessageByChannel",
                "parameters": [
                    {
                        "name": "channel",
                        "in": "path",
                        "description": "Name of the channel to get the message from",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Message content"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Message retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Channel not found or no message available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No message available"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unexpected error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}